update pythong tests
