+++ tests/integration/suite/alert_common.py
@@ -2,14 +2,31 @@ from flask import request
 from flask import Flask
 from threading import Thread
 from werkzeug.serving import make_server
+import socket
+
+
+def get_ip():
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    try:
+        # doesn't even have to be reachable
+        s.connect(('10.255.255.255', 1))
+        IP = s.getsockname()[0]
+    except Exception:
+        IP = '127.0.0.1'
+    finally:
+        s.close()
+    return IP
+
+
+LOCAL_IP = get_ip()
 
 
 class MockServer(Thread):
     def __init__(self, port=5000):
         self.port = port
         self.app = Flask(__name__)
-        self.server = make_server('127.0.0.1', self.port, self.app)
-        self.url = "http://127.0.0.1:%s" % self.port
+        self.server = make_server(LOCAL_IP, self.port, self.app)
+        self.url = "http://%s:%s" % (LOCAL_IP, self.port)
         self.thread = None
 
     def stop(self):
