+++ tests/integration/suite/conftest.py
@@ -4,7 +4,6 @@ import requests
 import time
 import urllib3
 import yaml
-import socket
 import subprocess
 import json
 import rancher
@@ -25,21 +24,8 @@ from .cluster_common import \
 urllib3.disable_warnings()
 
 
-def get_ip():
-    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    try:
-        # doesn't even have to be reachable
-        s.connect(('10.255.255.255', 1))
-        IP = s.getsockname()[0]
-    except Exception:
-        IP = '127.0.0.1'
-    finally:
-        s.close()
-    return IP
-
-
-IP = get_ip()
-SERVER_URL = os.environ.get('CATTLE_TEST_URL', 'https://' + IP + ':8443')
+IP = "localhost"
+SERVER_URL = os.environ.get('CATTLE_TEST_URL', 'https://' + IP + ':443')
 SERVER_PASSWORD = os.environ.get('RANCHER_SERVER_PASSWORD', 'admin')
 BASE_URL = SERVER_URL + '/v3'
 AUTH_URL = BASE_URL + '-public/localproviders/local?action=login'
